+++ python/tests/kat/test_ambassador.py
@@ -13,49 +13,51 @@ if letter_range not in ["all", "ah", "ip", "qz"]:
     sys.exit(1)
 
 if letter_range in ["all","ah"]:
-	import t_basics
-	import t_circuitbreaker
-	import t_cluster_tag
-	import t_consul
-	import t_cors
-	import t_envoy_logs
-	import t_error_response
-	import t_extauth
-	import t_grpc
-	import t_grpc_bridge
-	import t_grpc_stats
-	import t_grpc_web
-	import t_gzip
-	import t_headerrouting
-	import t_headerswithunderscoresaction
-	import t_hosts
+    import t_basics
+    import t_circuitbreaker
+    import t_cluster_tag
+    import t_consul
+    import t_cors
+    import t_envoy_logs
+    import t_error_response
+    import t_extauth
+    import t_grpc
+    import t_grpc_bridge
+    import t_grpc_stats
+    import t_grpc_web
+    import t_gzip
+    import t_headerrouting
+    import t_headerswithunderscoresaction
+    import t_hosts
 
 if letter_range in ["all","ip"]:
-	import t_ingress
-	import t_ip_allow_deny
-	import t_listeneridletimeout
-	import t_loadbalancer
-	import t_logservice
-	import t_lua_scripts
-	import t_max_req_header_kb
-	import t_no_ui
-    # t_plain include t_mappingtests and t_optiontests as imports
+    import t_ingress
+    import t_ip_allow_deny
+    import t_listeneridletimeout
+    import t_loadbalancer
+    import t_logservice
+    import t_lua_scripts
+    import t_max_req_header_kb
+    import t_no_ui
+    # mapping tests executed in the default namespace
+    import t_mappingtests_default
+    # t_plain include t_mappingtests_plain and t_optiontests as imports
     # these tests require each other and need to be executed as a set
-	import t_plain
+    import t_plain
 
 if letter_range in ["all","qz"]:
-	import t_queryparameter_routing
-	import t_ratelimit
-	import t_redirect
-	import t_regexrewrite_forwarding
-	import t_request_header
-	import t_retrypolicy
-	#import t_shadow
-	# t_stats has tests for statsd and dogstatsd. It's too flaky to run all the time.
-	# import t_stats
-	import t_tcpmapping
-	import t_tls
-	import t_tracing
+    import t_queryparameter_routing
+    import t_ratelimit
+    import t_redirect
+    import t_regexrewrite_forwarding
+    import t_request_header
+    import t_retrypolicy
+    #import t_shadow
+    # t_stats has tests for statsd and dogstatsd. It's too flaky to run all the time.
+    # import t_stats
+    import t_tcpmapping
+    import t_tls
+    import t_tracing
 
 # pytest will find this because Runner is a toplevel callable object in a file
 # that pytest is willing to look inside.
